Crate glib_win32

Rust GLib Win32 bindings

Rust bindings and wrappers for glib, part of gtk-rs-core.

GLib 2.56 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.83.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
glib = "0.13"
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

See Also

gio

License

__glib-win32 is available under the MIT License, please refer to it.

Re-exports

pub use glib_win32_sys as ffi;

Modules

prelude

Enums

OSType

Type of Windows edition to check for at run-time.

Functions

check_windows_version

command_line

error_message

ftruncate

getlocale

locale_filename_from_utf8

package_installation_directory

package_installation_subdirectory

windows_version