glib_win32

Function package_installation_directory_of_module

pub fn package_installation_directory_of_module(
    hmodule: HANDLE,
) -> Result<PathBuf, Error>

Available on Windows or docsrs only.

This function tries to determine the installation directory of a software package based on the location of a DLL of the software package.

@hmodule should be the handle of a loaded DLL or None. The function looks up the directory that DLL was loaded from. If @hmodule is NULL, the directory the main executable of the current process is looked up. If that directory’s last component is “bin” or “lib”, its parent directory is returned, otherwise the directory itself.

It thus makes sense to pass only the handle to a “public” DLL of a software package to this function, as such DLLs typically are known to be installed in a “bin” or occasionally “lib” subfolder of the installation folder. DLLs that are of the dynamically loaded module or plugin variety are often located in more private locations deeper down in the tree, from which it is impossible for GLib to deduce the root of the package installation.

The typical use case for this function is to have a DllMain() that saves the handle for the DLL. Then when code in the DLL needs to construct names of files in the installation tree it calls this function passing the DLL handle.

hmodule

The Win32 handle for a DLL loaded into the current process, or None

Returns

a string containing the guessed installation directory for the software package @hmodule is from. The string is in the GLib file name encoding, i.e. UTF-8. The return value should be freed with g_free() when not needed any longer. If the function fails None is returned.