glib_win32::prelude

Trait ParamSpecBuilderExt

pub trait ParamSpecBuilderExt<'a>: Sized {

    // Required methods
    fn set_nick(&mut self, nick: Option<&'a str>);
    fn set_blurb(&mut self, blurb: Option<&'a str>);
    fn set_flags(&mut self, flags: ParamFlags);
    fn current_flags(&self) -> ParamFlags;

    // Provided methods
    fn nick(self, nick: &'a str) -> Self { ... }
    fn blurb(self, blurb: &'a str) -> Self { ... }
    fn flags(self, flags: ParamFlags) -> Self { ... }
    fn read_only(self) -> Self { ... }
    fn write_only(self) -> Self { ... }
    fn readwrite(self) -> Self { ... }
    fn construct(self) -> Self { ... }
    fn construct_only(self) -> Self { ... }
    fn lax_validation(self) -> Self { ... }
    fn explicit_notify(self) -> Self { ... }
    fn deprecated(self) -> Self { ... }

}

A trait implemented by the various ParamSpec builder types.

It is useful for providing a builder pattern for ParamSpec defined outside of GLib like in GStreamer or GTK 4.

Required Methods

fn set_nick(&mut self, nick: Option<&'a str>)

Implementation detail.

fn set_blurb(&mut self, blurb: Option<&'a str>)

Implementation detail.

fn set_flags(&mut self, flags: ParamFlags)

Implementation detail.

fn current_flags(&self) -> ParamFlags

Implementation detail.

Provided Methods

fn nick(self, nick: &'a str) -> Self

By default, the nickname of its redirect target will be used if it has one. Otherwise, self.name will be used.

fn blurb(self, blurb: &'a str) -> Self

Default: None

fn flags(self, flags: ParamFlags) -> Self

Default: glib::ParamFlags::READWRITE

fn read_only(self) -> Self

Mark the property as read only and drops the READWRITE flag set by default.

fn write_only(self) -> Self

Mark the property as write only and drops the READWRITE flag set by default.

fn readwrite(self) -> Self

Mark the property as readwrite, it is the default value.

fn construct(self) -> Self

Mark the property as construct

fn construct_only(self) -> Self

Mark the property as construct only

fn lax_validation(self) -> Self

Mark the property as lax validation

fn explicit_notify(self) -> Self

Mark the property as explicit notify

fn deprecated(self) -> Self

Mark the property as deprecated

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.